Commit 11/14

+++ src/physicsserver/Machine.java
@@ -1,10 +1,10 @@
 package physicsserver;
 
 public class Machine {
-    public float distance;
+    public double distance;
     public String ip;
     
-    public Machine(String ip, float distance){
+    public Machine(String ip, double distance){
         this.ip = ip;
         this.distance = distance;
     }
